Optimize get_winner function
This commit will optimize this function by getting the value needed from
the comps (comp.get_sum('replica') or comp.get_total()) once, the first
time needed, saving that value into a variable, and then using that,
instead of calling the function everytime it is needed.

It also changes SaveWinner's behavior and purpuse, to what the name says
it will do. This function no longer calculates the winner and returns
it, it only saves the winner and lets the rest to the get_winner
function.