Validate input and add .5 input possibility
This commit will validate and customize all the input you give it.
If you put an space, it will translate it to '.5', and it won't allow
you to give invalid input.
This is applied on a function that is implemented in validation.ts, and
used on util.ts, just after creating the inputs.

To make this, I also had to make some changes on the API, more
specifically, on the SaveModes mutation and in the CompMode type.
Changed value_1 and 2 to a list of floats, and changed the type of
CompMode to a float instead of int.