Fix bug on mode inputs for computer and mobile
This commit fixes bugs like when an iOS user autocompleted a word, you
weren't able to delete that. It now uses inputMode decimal, so all
mobile devices use a numpad, which doesn't has autocompletion.

Android devices use a different name for the '.' in the numpad, and iOS
devices used a comma instead, so created decimals arr to compare the key
with that arr to check for decimal key use that functionality.

It also now checks for keys that are allowed by default with validInputs
arr (this only works for computer devices).

Renamed validate_inputs to validateCompInputs

v0.1.0