Add space/newline between rhyme and extra inptus
This commit will add space between rhyme and extra inputs on medium to
large screens. On small screens, the extra-inputs will be on a new line.
This also makes input-(left, right, middle) classes to go on the last
rhyme-inputs.

This makes a series of changes on different files:
- polls.scss
  - Make change-grid span input-middle through 3 cols.
  - Make align available for input-extra-(side)
  - Use a for loop to make .input-(side) and .input-extra-(side)
  - .input now centers and moves to the top everything
- util.ts
  - Check for .input-(side) on the last 2 rhyme inputs
  - Add .input-extra-(side) on the antepenultimate and last inputs (the
  extra inputs)

v0.2.0